+++ R/cl.R
@@ -89,27 +89,41 @@ cl_lexicon_size <- function(corpus, p_attribute, registry = Sys.getenv("CORPUS_R
 #' txt <- cl_cpos2str("REUTERS", p_attribute = "word", cpos = cpos_struc[1]:cpos_struc[2])
 #' fulltext <- paste(txt, collapse = " ")
 cl_cpos2struc <- function(corpus, s_attribute, cpos, registry = Sys.getenv("CORPUS_REGISTRY")){
-  retval <- .cl_cpos2struc(corpus = corpus, s_attribute = s_attribute, cpos = cpos, registry = registry)
-  retval
+  check_registry(registry)
+  check_corpus(corpus, registry)
+  check_s_attribute(corpus = corpus, registry = registry, s_attribute = s_attribute)
+  .cl_cpos2struc(corpus = corpus, s_attribute = s_attribute, cpos = cpos, registry = registry)
 }
 
 #' @rdname s_attributes
 cl_struc2cpos <- function(corpus, s_attribute, registry = Sys.getenv("CORPUS_REGISTRY"), struc){
+  check_registry(registry)
+  check_corpus(corpus, registry)
+  check_s_attribute(corpus = corpus, registry = registry, s_attribute = s_attribute)
   .cl_struc2cpos(corpus = corpus, s_attribute = s_attribute, registry = registry, struc = struc)
 }
 
 #' @rdname s_attributes
 cl_struc2str <- function(corpus, s_attribute, struc, registry = Sys.getenv("CORPUS_REGISTRY")){
+  check_registry(registry)
+  check_corpus(corpus, registry)
+  check_s_attribute(corpus = corpus, registry = registry, s_attribute = s_attribute)
   .cl_struc2str(corpus = corpus, s_attribute = s_attribute, struc = struc, registry = registry)
 }
 
 #' @rdname s_attributes
 cl_cpos2lbound <- function(corpus, s_attribute, cpos, registry = Sys.getenv("CORPUS_REGISTRY")){
+  check_registry(registry)
+  check_corpus(corpus, registry)
+  check_s_attribute(corpus = corpus, registry = registry, s_attribute = s_attribute)
   .cl_cpos2lbound(corpus = corpus, s_attribute = s_attribute, cpos = cpos, registry = registry)
 }
 
 #' @rdname s_attributes
 cl_cpos2rbound <- function(corpus, s_attribute, cpos, registry = Sys.getenv("CORPUS_REGISTRY")){
+  check_registry(registry)
+  check_corpus(corpus, registry)
+  check_s_attribute(corpus = corpus, registry = registry, s_attribute = s_attribute)
   .cl_cpos2rbound(corpus = corpus, s_attribute = s_attribute, cpos = cpos, registry = registry)
 }
 
@@ -191,51 +205,53 @@ cl_cpos2rbound <- function(corpus, s_attribute, cpos, registry = Sys.getenv("COR
 #' )
 #' 
 cl_cpos2str <- function(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"), cpos){
-  .check_registry(registry)
-  .check_corpus(corpus, registry)
+  check_registry(registry)
+  check_corpus(corpus, registry)
   .cl_cpos2str(corpus = corpus, p_attribute = p_attribute, registry = registry, cpos = cpos)
 }
 
 #' @rdname p_attributes
 cl_cpos2id <- function(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"), cpos){
-  .check_registry(registry)
-  .check_corpus(corpus, registry)
+  check_registry(registry)
+  check_corpus(corpus, registry)
   .cl_cpos2id(corpus = corpus, p_attribute = p_attribute, registry = registry, cpos = cpos)
 }
 
 #' @rdname p_attributes
 cl_id2str <- function(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"), id){
-  .check_registry(registry)
-  .check_corpus(corpus, registry)
+  check_registry(registry)
+  check_corpus(corpus, registry)
   .cl_id2str(corpus = corpus, p_attribute = p_attribute, registry = registry, id = id)
 }
 
 #' @rdname p_attributes
 cl_regex2id <- function(corpus, p_attribute, regex, registry = Sys.getenv("CORPUS_REGISTRY")){
-  .check_registry(registry)
-  .check_corpus(corpus, registry)
+  check_registry(registry)
+  check_corpus(corpus, registry)
   .cl_regex2id(corpus = corpus, p_attribute = p_attribute, regex = regex, registry = registry)
 }
 
 #' @rdname p_attributes
 cl_str2id <- function(corpus, p_attribute, str, registry = Sys.getenv("CORPUS_REGISTRY")){
-  .check_registry(registry)
-  .check_corpus(corpus, registry)
+  check_registry(registry)
+  check_corpus(corpus, registry)
   .cl_str2id(corpus = corpus, p_attribute = p_attribute, str = str, registry = registry)
 }
 
 #' @rdname p_attributes
 cl_id2freq <- function(corpus, p_attribute, id, registry = Sys.getenv("CORPUS_REGISTRY")){
-  .check_registry(registry)
-  .check_corpus(corpus, registry)
+  check_registry(registry)
+  check_corpus(corpus, registry)
+  check_p_attribute(p_attribute = p_attribute, corpus = corpus, registry = registry)
   .cl_id2freq(corpus = corpus, p_attribute = p_attribute, id = id, registry = registry)
 }
 
 
 #' @rdname p_attributes
 cl_id2cpos <- function(corpus, p_attribute, id, registry = Sys.getenv("CORPUS_REGISTRY")){
-  .check_registry(registry)
-  .check_corpus(corpus, registry)
+  check_registry(registry)
+  check_corpus(corpus, registry)
+  check_p_attribute(p_attribute = p_attribute, corpus = corpus, registry = registry)
   .cl_id2cpos(corpus = corpus, p_attribute = p_attribute, id = id, registry = registry)
 }